chore(logic): Runner should be removed before termination
The EC2 instance should be marked as unschedulable by Github Actions
before termination to make sure that Github Actions dosen't schedule any
jobs to it while it's being terminated.

Cherry-pick of:

- https://github.com/machulav/ec2-github-runner/pull/161

This seems to be solving the same problem as the `--ephemeral` argument
in the preceding commit but I'm deep enough in this yak shave that I
don't want to debug which one works better right now.